#340 & #244. Isolating i18n strings, especially those in value=" parameter. Correcting some indentations. part 2/5

+++ handlers/page/delete_referrer.php
@@ -10,26 +10,28 @@
  * 
  * @uses	Wakka::Href()
  * @uses	Wakka::Query()
- * @uses	Wakka::redirect()
+ * @uses	Wakka::Redirect()
  */
 
-if ($spam_link = $_REQUEST["spam_link"])
+if (isset($_REQUEST['spam_link']))
 {
-	$parsed_url = parse_url($spam_link);
-	$spammer = $parsed_url["host"];
-
-} else {
-	$spammer = $_REQUEST["spam_site"];
+	$parsed_url = parse_url($_REQUEST['spam_link']);
+	$spammer = isset($parsed_url['host']) ? $parsed_url['host'] : '';
+}
+elseif (isset($_REQUEST['spam_site']))
+{
+	$spammer = $_REQUEST['spam_site'];
 }
 
-if ($spammer) {
+if (isset($spammer) && $spammer)
+{
 	$this->Query("DELETE FROM ".$this->config["table_prefix"]."referrers WHERE referrer like '%".mysql_real_escape_string($spammer)."%'");
 	if (!$already_blacklisted = $this->LoadSingle("select * from ".$this->config["table_prefix"]."referrer_blacklist WHERE spammer = '".mysql_real_escape_string($spammer)."'"))
 		$this->Query("INSERT INTO ".$this->config["table_prefix"]."referrer_blacklist SET spammer = '".mysql_real_escape_string($spammer)."'");
 }
 
 // Redirect to last page
-$redirect = $_REQUEST["redirect"];
-$this->redirect($this->Href($redirect));
+$redirect = isset($_REQUEST['redirect']) ? $_REQUEST['redirect'] : '';
+$this->Redirect($this->Href($redirect));
 
 ?>
